a generated corpus is a GATE only when its oracle is independent AND its open set is written down
`tests/exhaustive` landed last commit RED — 274 findings — and a red generator is
a document, not a gate. It is a gate now: two ctest tests, both green, and each
one bites for every class it was built to catch. What made green possible was
closing the half of the open root that produced WRONG ANSWERS, because a wrong
answer is the one thing a ledger may never hold.

── THE ARC: a range-pattern bound is as wide as the SCRUTINEE ────────────────

`PatRange { lo: i64, hi: i64 }` was the scrutinee's width for exactly ten of the
sixteen integer types. An `i128`/`u128` bound was TRUNCATED to its low 64 bits
and still compiled, so the emitted test covered a different range and answered
confidently: `170141183460469231731687303715884105726..=…727` matched nothing.
Widened end to end — `parse_int_literal_u128` in sema, two mirror halves
(`LO_HI`/`HI_HI`, written only when they are not the sign-extension of the low
half, so every sub-128-bit pattern produces byte-identical mirror bytes), and a
full-width `arith.constant` in `emit_range_test`. There is no `int64_t lo()` on
`PatRangeView` any more: a narrowing read is what the defect was.

The bound is now a PATTERN of the scrutinee's width compared in the scrutinee's
signedness, rather than a raw `int64_t` that conflated "the value 2^63" with
"the value -2^63". The pre-existing reinterpretation (a magnitude between |max|
and the unsigned ceiling on a SIGNED scrutinee spells its two's complement) is
preserved and now written down instead of falling out of `parse_int_literal`.

MEASURED: pattern family 112 findings -> 0, including all 24 wrong answers.
Silent wrong answers over the whole 13 508-case product: 0.

── THE GATE ─────────────────────────────────────────────────────────────────

  logos_26_exhaustive_smoke   86 progs / 11 316 cases / 34 s — EVERY level of
      test-levels.sh (it is not a corpus member, so the L1-L3 samplers cannot
      reach it; it is named explicitly there), and L4.
  logos_26_exhaustive_full   246 progs / 13 508 cases / 153 s — L4.

`smoke` is DECLARED, not sampled: cast/cmp/arith/pattern/poison IN FULL (they
cost 11 s together — thousands of cases per compile), plus a DIAGONAL through
deem, one program per payload type walking the nine (field-position x
caller-context) combinations, so all 20 types and all 3 positions and all 3
contexts appear in 20 programs of 180. What smoke does NOT run is stated rather
than left to be inferred: the other 160 deem programs, 2 192 cases, at L4.

Each tier asserts three things, in order:
  1. the CORPUS DIGEST matches `corpus.<tier>.sha256` — a run against a drifted
     generator is a run about nothing. The corpus text is not checked in (a
     generated file that could be edited stops being a spec); the digest is.
  2. ZERO wrong answers, absolutely, no ledger and no tier exemption.
  3. the REFUSAL set is EXACTLY `refusals.ledger`, three directions: unlisted
     refusal = new defect; listed refusal that now compiles = the arc landed and
     the ledger is stale; line naming no case at all = asserts nothing. Adding a
     line and walking away cannot make this green.

The 162 remaining refusals are one root, named in the ledger: an integer
literal's value is carried in 64 SIGNED bits, from `wstr_decode_i64` in the WQL
surface tokenizer (duplicated in el_parser and trama_parser) to codegen.logos:985.
108 need a 64-bit UNSIGNED carriage, 54 a 128-bit one. One arc, reported rather
than half-done — and ledgerable ONLY because the refusal is loud.

── PROVEN TO BITE, ctest exit 8 each, tree restored byte-identical after ─────

  revert attach_target_data_layout (8ba3c764) -> 34-35 wrong answers, naming
      deem.{i8,u8,i16,u16,i24,u24,i32,u32,i128}.{0,1,2}.{bare,padbefore,padafter}
  revert `uns` in emit_range_test (ac81ba99) -> 96, naming pat.u8/u16 at all
      four of {match,iflet,letelse,whilelet}
  revert key_ord_frag (519a181d)             -> 4, naming deem.f64.2.padbefore
      and deem.f32.2.padafter, order_asc/order_desc
  revert this commit's 128-bit PatRange half -> 56, naming pat.i128 at all four
      sites over -2^126..=2^126-1

── AND IT CANNOT BE FOOLED THE WAY THE FIXTURE WAS ──────────────────────────

With key_ord_frag reverted, `tests/logos/pass/wql_join_order_key_fidelity_e2e`
AS IT SHIPPED (519a181d^) still exits 0 — its oracle was an insertion sort in
the same file using the same partial `>`, so it halted at the NaN in exactly the
same place and compared the implementation with itself. On that same compiler
the enumerator says:

    NOT ascending at 1.0 > -inf; key sequence
      [-1.0, 0.0, -0.0, 1.0, -inf, -1.797e+308, 0.5, 1.797e+308, inf]

because `_order_check` reads monotonicity OUT OF the answer and never sorts
anything. That is the whole difference between this gate and what it replaces.

Gates observed: build clean; L2 1859/1859 + the smoke tier (11 316 generated
cases); conuco 67/0; ABI PRESERVING; L4 3114/3114. ctest 6796 -> 6798, exactly
the two tests added. Corpus digests regenerate byte-identical.

+++ include/logos/compiler/lir_mirror.hpp
@@ -289,7 +289,7 @@ const uint8_t* lir_mirror_emit_pat_wild         (lir::LProgram& prog, std::strin
 const uint8_t* lir_mirror_emit_pat_variant_data (lir::LProgram& prog, std::string_view enum_name, std::string_view variant, int64_t disc, const std::vector<std::string>& bindings, const std::vector<TypeRef>& binding_types, const std::vector<uint32_t>& bind_slots = {});
 const uint8_t* lir_mirror_emit_pat_or           (lir::LProgram& prog, const std::vector<lir::Pattern>& alts);
 const uint8_t* lir_mirror_emit_pat_tuple        (lir::LProgram& prog, const std::vector<std::string>& bindings, const std::vector<TypeRef>& binding_types, const std::vector<lir::Pattern>& subs, const std::vector<uint32_t>& bind_slots = {});
-const uint8_t* lir_mirror_emit_pat_range        (lir::LProgram& prog, int64_t lo, int64_t hi);
+const uint8_t* lir_mirror_emit_pat_range        (lir::LProgram& prog, __int128 lo, __int128 hi);
 const uint8_t* lir_mirror_emit_pat_struct       (lir::LProgram& prog, std::string_view struct_name, const std::vector<lir::PatFieldBinding>& fields, bool has_rest);
 const uint8_t* lir_mirror_emit_pat_slice        (lir::LProgram& prog, const std::vector<lir::Pattern>& prefix, const std::vector<lir::Pattern>& rest, const std::vector<lir::Pattern>& suffix);
 const uint8_t* lir_mirror_emit_pat_at           (lir::LProgram& prog, std::string_view name, const std::vector<lir::Pattern>& sub, TypeRef type, uint32_t slot = 0xFFFFFFFFu);
